# Optimize String write