--- a/bson/src/main/org/bson/ByteBuf.java
+++ b/bson/src/main/org/bson/ByteBuf.java
@@ -136,6 +136,8 @@ public interface ByteBuf  {
      */
     byte[] array();
 
+    boolean hasArray();
+
     /**
      * Returns this buffer's limit.
      *