--- a/bson/src/main/org/bson/ByteBufNIO.java
+++ b/bson/src/main/org/bson/ByteBufNIO.java
@@ -109,6 +109,11 @@ public class ByteBufNIO implements ByteBuf {
     }
 
     @Override
+    public boolean hasArray() {
+        return buf.hasArray();
+    }
+
+    @Override
     public int limit() {
         return buf.limit();
     }