--- a/driver-core/src/main/com/mongodb/internal/connection/PowerOfTwoBufferPool.java
+++ b/driver-core/src/main/com/mongodb/internal/connection/PowerOfTwoBufferPool.java
@@ -156,7 +156,7 @@ public class PowerOfTwoBufferPool implements BufferProvider {
         return v;
     }
 
-    public class PooledByteBufNIO extends ByteBufNIO {
+    private class PooledByteBufNIO extends ByteBufNIO {
 
         PooledByteBufNIO(final ByteBuffer buf) {
             super(buf);