--- a/driver-core/src/main/com/mongodb/internal/connection/netty/NettyByteBuf.java
+++ b/driver-core/src/main/com/mongodb/internal/connection/netty/NettyByteBuf.java
@@ -101,6 +101,11 @@ public final class NettyByteBuf implements ByteBuf {
     }
 
     @Override
+    public boolean hasArray() {
+        return proxied.hasArray();
+    }
+
+    @Override
     public int limit() {
         if (isWriting) {
             return proxied.writerIndex() + remaining();