--- a/driver-core/src/main/com/mongodb/internal/connection/netty/NettyByteBuf.java
+++ b/driver-core/src/main/com/mongodb/internal/connection/netty/NettyByteBuf.java
@@ -106,6 +106,11 @@ public final class NettyByteBuf implements ByteBuf {
     }
 
     @Override
+    public int arrayOffset() {
+        return proxied.arrayOffset();
+    }
+
+    @Override
     public int limit() {
         if (isWriting) {
             return proxied.writerIndex() + remaining();