@franz1981 2025-03-18
That's related to https://github.com/mongodb/mongo-java-driver/pull/1629 🙏