@vbabanin 2025-04-17
I've added additional tests for code points represented by surrogate pairs:
https://github.com/mongodb/mongo-java-driver/pull/1651/commits/dd5fe4de84e04e5b155080d4a759e4363671b4f8

Adding a second reviewer to look at just this commit -  the rest has already been reviewed by @rozza.